+++ test/test_kojo/modules/charlie/methodA.js
@@ -1,7 +1,7 @@
 module.exports = async function (arg) {
 
-    const {plant, logger} = this;
-    const variable = plant.get('variable');
+    const {kojo, logger} = this;
+    const variable = kojo.get('variable');
     if (typeof logger.info !== 'function') throw new Error('Logger inaccessible');
     if(!arg) throw new Error('No arguments defined (error log test)');
     return arg * variable;